Promisify the AuthApp.
The code looks MUCH clearer now.  Not all of my tests pass, though,
because I'm in the process of removing the use of EventEmitter in the
app.

Also, while I was at it, I decided to change the way I approach loading
fixtures.  I'll soon be able to get rid of the "dbImport" monstrosity.