Replace config module with ini parser.
Instead of using the config module, we now read an ini file when we run
charonauth and pass the configuration values down through the processes
using process messages.

Also, while I was at it, promisify the webapp as well, to bring it in
line with the authapp.  Also break out logging into its own module and
config section.