Introduce verbose logging.
Pass through a "verbose logging" flag either from the command line or
from the ini file that allows us to print extra logging data to the
console.  Currently, the logger is created in the scope of the master
and workers, and the worker logger is injected into the app objects
during construction.  Perhaps top-level logging will come in the
future...

While I was at it, I fixed-up the auth and webmasters to automatically
restart if killed, prevented them from repsawn-spamming if the workers
failed to construct, and broke out reCAPTCHA verification to its own
module.