Start work on Action tracking.
Currently we track when a user logs in and authenticates, and use that
data to display the Action log as well as the "Last Played" statistic on
the user's page.

Also, did an anti-zalgo work on the authentication app by covering all
route functionality with a promise, not just from the first async
function on.